fixed various code smells
