+++ frontend/src/libs/import-export/index.js
@@ -6,5 +6,4 @@ export * from "./normalize-import";
 export * from "./validate";
 export * from "./prepare-spreadsheet";
 export * from "./prepare-json";
-export * from "./parse-spreadsheet";
 export * from "./normalize-ddah-import";
