+++ frontend/src/tests/import-export-data/applicants_missing_required_keys.csv
@@ -1,5 +1,5 @@
 ﻿Last Name,First Name,UTORid,Student Number,email,Phone, age
-Doe,John,,10000000,goofy-duck@donald.com,4166666666,
+Doe,John,,="10000000",goofy-duck@donald.com,="4166666666",
 Weasley,Ron,weasleyr,,,,
-Potter,哈利,potterh,999666999,harry@potter.com,41888888888,
-Smith,Bethany,smithb,131382748,smithb@mail.utoronto.ca,41999999999,33
+Potter,哈利,potterh,="999666999",harry@potter.com,="41888888888",
+Smith,Bethany,smithb,="131382748",smithb@mail.utoronto.ca,="41999999999",33
