+++ frontend/src/tests/import-export.test.js
@@ -2,7 +2,7 @@
  * @jest-environment node
  */
 /* eslint-env node */
-import { describe, it, expect } from "./utils";
+import { describe, it, expect, parseSpreadsheet } from "./utils";
 import {
     instructorData,
     applicantData,
@@ -26,9 +26,20 @@ import {
     prepareSpreadsheet,
     SpreadsheetRowMapper,
     normalizeImport,
-    parseSpreadsheet,
     normalizeDdahImports,
 } from "../libs/import-export";
+import {
+    importInstructorContext,
+    importApplicantContext,
+    importPositionContext,
+    importPositionContextNoSmithHenry,
+    importAssignmentContext,
+    importAssignmentContextNoPosition,
+    importAssignmentContextNoApplicant,
+    importDdahApplicantContext,
+    importDdahContext,
+    importDdahContextNoAssignment,
+} from "./import-export-data/import-context";
 
 // create a shim for native File object for export round trip test
 function File(fileBits, fileName, options) {
@@ -220,23 +231,7 @@ describe("Import/export library functionality", () => {
         // compute the difference between existing instructors
         const instructorsDiff = diffImport.instructors(
             normalizedSpreadsheetInstructors,
-            {
-                instructors: [
-                    // instructor to be modified
-                    {
-                        first_name: "Henry",
-                        utorid: "smithh",
-                        email: "OLD@utoronto.ca",
-                    },
-                    // instructor to be duplicated
-                    {
-                        email: "gordon.smith@utoronto.ca",
-                        first_name: "戈登",
-                        last_name: "Smith",
-                        utorid: "smithhg",
-                    },
-                ],
-            }
+            importInstructorContext
         );
         expect(instructorsDiff).toMatchSnapshot();
     });
@@ -275,27 +270,7 @@ describe("Import/export library functionality", () => {
         // compute the difference between existing applicants
         const applicantsDiff = diffImport.applicants(
             normalizedSpreadsheetApplicants,
-            {
-                applicants: [
-                    // applicant to be modified
-                    {
-                        first_name: "John",
-                        last_name: "Doe",
-                        utorid: "johnd",
-                        email: "goofy-duck@donald.com",
-                        student_number: "OLD10000000",
-                    },
-                    // applicant to be duplicated
-                    {
-                        first_name: "哈利",
-                        last_name: "Potter",
-                        utorid: "potterh",
-                        email: "harry@potter.com",
-                        student_number: "999666999",
-                        phone: "41888888888",
-                    },
-                ],
-            }
+            importApplicantContext
         );
         expect(applicantsDiff).toMatchSnapshot();
     });
@@ -340,112 +315,26 @@ it("Import Positions from JSON/CSV/XLSX", () => {
             positionSchema
         ).toThrow(Error)
     );
+    // make sure the instructors are valid instructors
     for (const item of normalizedSpreadsheetPositions) {
         item.instructors = diffImport
             .instructorsListFromField(item.instructors || [], {
-                instructors: [
-                    {
-                        first_name: "Henry",
-                        last_name: "Smith",
-                        utorid: "smithh",
-                    },
-                    {
-                        first_name: "Emily",
-                        last_name: "Garcia",
-                        utorid: "garciae",
-                    },
-                ],
+                instructors: importPositionContext.instructors,
             })
             .map((x) => x.utorid);
     }
     // compute the difference with existing positions
-    const positionsDiff = diffImport.positions(normalizedSpreadsheetPositions, {
-        positions: [
-            {
-                position_code: "MAT136H1F",
-                position_title: "代数",
-                hours_per_assignment: 70,
-                start_date: "2020-02-10T00:00:00.000Z",
-                end_date: "2020-12-31T00:00:00.000Z",
-                instructors: [
-                    {
-                        first_name: "Henry",
-                        last_name: "Smith",
-                        utorid: "smithh",
-                    },
-                    {
-                        first_name: "Emily",
-                        last_name: "Garcia",
-                        utorid: "garciae",
-                    },
-                ],
-                contract_template: {
-                    template_name: "Regular",
-                },
-                duties: "",
-                qualifications: "",
-            },
-            {
-                position_code: "CSC494",
-                position_title: "Capstone Project",
-                hours_per_assignment: 20,
-                start_date: "2020-02-10T00:00:00.000Z",
-                end_date: "2020-12-31T00:00:00.000Z",
-                instructors: [
-                    {
-                        first_name: "Henry",
-                        last_name: "Smith",
-                        utorid: "smithh",
-                    },
-                    {
-                        first_name: "Emily",
-                        last_name: "Garcia",
-                        utorid: "garciae",
-                    },
-                ],
-                contract_template: {
-                    template_name: "Regular",
-                },
-                duties: "",
-                qualifications: "",
-            },
-        ],
-        instructors: [
-            {
-                first_name: "Henry",
-                last_name: "Smith",
-                utorid: "smithh",
-            },
-            {
-                first_name: "Emily",
-                last_name: "Garcia",
-                utorid: "garciae",
-            },
-        ],
-        contractTemplates: [
-            {
-                template_name: "Regular",
-            },
-        ],
-    });
+    const positionsDiff = diffImport.positions(
+        normalizedSpreadsheetPositions,
+        importPositionContext
+    );
     expect(positionsDiff).toMatchSnapshot();
     // import positions assigned to invalid instructors should throw error
     expect(() =>
-        diffImport.positions(normalizedSpreadsheetPositions, {
-            positions: [],
-            instructors: [
-                {
-                    first_name: "Emily",
-                    last_name: "Garcia",
-                    utorid: "garciae",
-                },
-            ],
-            contractTemplates: [
-                {
-                    template_name: "Regular",
-                },
-            ],
-        })
+        diffImport.positions(
+            normalizedSpreadsheetPositions,
+            importPositionContextNoSmithHenry
+        )
     ).toThrow(Error);
 });
 
@@ -491,150 +380,24 @@ it("Import Assignments from JSON/CSV/XLSX", () => {
         ).toThrow(Error)
     );
     // Compute the difference with existing assignments
-    const assignmentsDiff = diffImport.assignments(normalizedJsonAssignments, {
-        assignments: [
-            {
-                applicant: {
-                    first_name: "Harry",
-                    last_name: "Potter",
-                    email: "a@a.com",
-                    utorid: "potterh",
-                    phone: "41666666666",
-                    student_number: "1000000000",
-                },
-                position: {
-                    position_code: "CSC494",
-                    position_title: "Capstone Project",
-                    hours_per_assignment: 70,
-                    start_date: "2020-12-10T00:00:00.000Z",
-                    end_date: "2021-12-10T00:00:00.000Z",
-                    duties: "mark assignments",
-                    qualifications: "3 300-lvl CSC courses",
-                    ad_hours_per_assignment: null,
-                    ad_num_assignments: null,
-                    ad_open_date: null,
-                    ad_close_date: null,
-                    desired_num_assignments: 20,
-                    current_enrollment: 400,
-                    current_waitlisted: 100,
-                    instructors: [],
-                    contract_template: {
-                        template_name: "Regular",
-                    },
-                },
-                start_date: "2020-12-10T00:00:00.000Z",
-                end_date: "2021-12-10T00:00:00.000Z",
-                contract_override_pdf: null,
-                hours: 80,
-                active_offer_status: null,
-                active_offer_recent_activity_date: null,
-                // more than 2 wage_chunks
-                wage_chunks: [
-                    {
-                        hours: 30,
-                        rate: 50,
-                        start_date: "2020-12-10T00:00:00.000Z",
-                        end_date: "2020-12-31T00:00:00.000Z",
-                    },
-                    {
-                        hours: 30,
-                        rate: 50,
-                        start_date: "2021-01-01T00:00:00.000Z",
-                        end_date: "2021-06-30T00:00:00.000Z",
-                    },
-                    {
-                        hours: 20,
-                        rate: 50,
-                        start_date: "2021-07-01T00:00:00.000Z",
-                        end_date: "2021-12-10T00:00:00.000Z",
-                    },
-                ],
-            },
-        ],
-        positions: [
-            {
-                position_code: "CSC494",
-                position_title: "Capstone Project",
-                hours_per_assignment: 70,
-                start_date: "2020-12-10T00:00:00.000Z",
-                end_date: "2021-12-10T00:00:00.000Z",
-                duties: "mark assignments",
-                qualifications: "3 300-lvl CSC courses",
-                ad_hours_per_assignment: null,
-                ad_num_assignments: null,
-                ad_open_date: null,
-                ad_close_date: null,
-                desired_num_assignments: 20,
-                current_enrollment: 400,
-                current_waitlisted: 100,
-                instructors: [],
-                contract_template: {
-                    template_name: "Regular",
-                },
-            },
-        ],
-        applicants: [
-            {
-                first_name: "Harry",
-                last_name: "Potter",
-                email: "a@a.com",
-                utorid: "potterh",
-                phone: "41666666666",
-                student_number: "1000000000",
-            },
-            { first_name: "Ron", last_name: "Weasley", utorid: "weasleyr" },
-        ],
-        session: { rate: 50, rate1: 50, rate2: 50, rate3: 50 },
-    });
+    const assignmentsDiff = diffImport.assignments(
+        normalizedJsonAssignments,
+        importAssignmentContext
+    );
     expect(assignmentsDiff).toMatchSnapshot();
     // import assignments with invalid applicant should throw error
     expect(() =>
-        diffImport.assignments(normalizedJsonAssignments, {
-            assignments: [],
-            positions: [
-                {
-                    position_code: "CSC494",
-                    position_title: "Capstone Project",
-                    hours_per_assignment: 70,
-                    start_date: "2020-12-10T00:00:00.000Z",
-                    end_date: "2021-12-10T00:00:00.000Z",
-                    duties: "mark assignments",
-                    qualifications: "3 300-lvl CSC courses",
-                    ad_hours_per_assignment: null,
-                    ad_num_assignments: null,
-                    ad_open_date: null,
-                    ad_close_date: null,
-                    desired_num_assignments: 20,
-                    current_enrollment: 400,
-                    current_waitlisted: 100,
-                    instructors: [],
-                    contract_template: {
-                        template_name: "Regular",
-                    },
-                },
-            ],
-            applicants: [],
-            session: { rate: 50, rate1: 50, rate2: 50, rate3: 50 },
-        })
+        diffImport.assignments(
+            normalizedJsonAssignments,
+            importAssignmentContextNoApplicant
+        )
     ).toThrow(Error);
     // import assignments with invalid position should throw error
     expect(() =>
-        diffImport.assignments(normalizedJsonAssignments, {
-            assignments: [],
-            positions: [],
-            applicants: [
-                {
-                    first_name: "Harry",
-                    last_name: "Potter",
-                    email: "a@a.com",
-                    utorid: "potterh",
-                    phone: "41666666666",
-                    student_number: "1000000000",
-                },
-                { first_name: "Ron", last_name: "Weasley", utorid: "weasleyr" },
-            ],
-            session: { rate: 50, rate1: 50, rate2: 50, rate3: 50 },
-        })
+        diffImport.assignments(
+            normalizedJsonAssignments,
+            importAssignmentContextNoPosition
+        )
     ).toThrow(Error);
 });
 
@@ -645,16 +408,7 @@ it("Import Ddahs from JSON/CSV/XLSX", () => {
             fileType: "spreadsheet",
             data: parseSpreadsheet("ddahs_correct.xlsx"),
         },
-        [
-            {
-                first_name: "Hanna",
-                last_name: "Wilson",
-                email: "wilsonh@mail.utoronto.ca",
-                utorid: "wilsonh",
-                phone: "41666666666",
-                student_number: "1000000000",
-            },
-        ]
+        importDdahApplicantContext
     );
     expect(normalizedSpreadsheetDdahs).toMatchSnapshot();
     // import ddahs with invalid applicant should throw error
@@ -664,16 +418,7 @@ it("Import Ddahs from JSON/CSV/XLSX", () => {
                 fileType: "spreadsheet",
                 data: parseSpreadsheet("ddahs_invalid_applicant.xlsx"),
             },
-            [
-                {
-                    first_name: "Hanna",
-                    last_name: "Wilson",
-                    email: "wilsonh@mail.utoronto.ca",
-                    utorid: "wilsonh",
-                    phone: "41666666666",
-                    student_number: "1000000000",
-                },
-            ]
+            importDdahApplicantContext
         )
     ).toThrow(Error);
     // import correct ddahs from JSON
@@ -686,121 +431,10 @@ it("Import Ddahs from JSON/CSV/XLSX", () => {
     );
     expect(normalizedJsonDdahs).toMatchSnapshot();
     // Compute the difference with existing ddahs
-    const ddahsDiff = diffImport.ddahs(normalizedJsonDdahs, {
-        ddahs: [
-            {
-                assignment: {
-                    applicant: {
-                        first_name: "Harry",
-                        last_name: "Potter",
-                        email: "a@a.com",
-                        utorid: "potterh",
-                        phone: "41666666666",
-                        student_number: "1000000000",
-                    },
-                    position: {
-                        position_code: "CSC135H1F",
-                        hours_per_assignment: 70,
-                        start_date: "2020-12-10T00:00:00.000Z",
-                        end_date: "2021-12-10T00:00:00.000Z",
-                        duties: "mark assignments",
-                        qualifications: "3 300-lvl CSC courses",
-                        instructors: [],
-                        contract_template: {
-                            template_name: "Regular",
-                        },
-                    },
-                    start_date: "2020-12-10T00:00:00.000Z",
-                    end_date: "2021-12-10T00:00:00.000Z",
-                    contract_override_pdf: null,
-                    hours: 80,
-                    active_offer_status: null,
-                    active_offer_recent_activity_date: null,
-                    // more than 2 wage_chunks
-                    wage_chunks: [
-                        {
-                            hours: 30,
-                            rate: 50,
-                            start_date: "2020-12-10T00:00:00.000Z",
-                            end_date: "2020-12-31T00:00:00.000Z",
-                        },
-                        {
-                            hours: 30,
-                            rate: 50,
-                            start_date: "2021-01-01T00:00:00.000Z",
-                            end_date: "2021-06-30T00:00:00.000Z",
-                        },
-                        {
-                            hours: 20,
-                            rate: 50,
-                            start_date: "2021-07-01T00:00:00.000Z",
-                            end_date: "2021-12-10T00:00:00.000Z",
-                        },
-                    ],
-                },
-                duties: [
-                    {
-                        description: "Initial training",
-                        hours: 80,
-                    },
-                    {
-                        description: "Marking the midterm",
-                        hours: 50,
-                    },
-                ],
-            },
-        ],
-        assignments: [
-            {
-                applicant: {
-                    first_name: "Harry",
-                    last_name: "Potter",
-                    email: "a@a.com",
-                    utorid: "potterh",
-                    phone: "41666666666",
-                    student_number: "1000000000",
-                },
-                position: {
-                    position_code: "CSC135H1F",
-                    hours_per_assignment: 70,
-                    start_date: "2020-12-10T00:00:00.000Z",
-                    end_date: "2021-12-10T00:00:00.000Z",
-                    duties: "mark assignments",
-                    qualifications: "3 300-lvl CSC courses",
-                    instructors: [],
-                    contract_template: {
-                        template_name: "Regular",
-                    },
-                },
-                start_date: "2020-12-10T00:00:00.000Z",
-                end_date: "2021-12-10T00:00:00.000Z",
-                contract_override_pdf: null,
-                hours: 80,
-                active_offer_status: null,
-                active_offer_recent_activity_date: null,
-                // more than 2 wage_chunks
-                wage_chunks: [
-                    {
-                        hours: 30,
-                        rate: 50,
-                        start_date: "2020-12-10T00:00:00.000Z",
-                        end_date: "2020-12-31T00:00:00.000Z",
-                    },
-                    {
-                        hours: 30,
-                        rate: 50,
-                        start_date: "2021-01-01T00:00:00.000Z",
-                        end_date: "2021-06-30T00:00:00.000Z",
-                    },
-                    {
-                        hours: 20,
-                        rate: 50,
-                        start_date: "2021-07-01T00:00:00.000Z",
-                        end_date: "2021-12-10T00:00:00.000Z",
-                    },
-                ],
-            },
-        ],
-    });
+    const ddahsDiff = diffImport.ddahs(normalizedJsonDdahs, importDdahContext);
     expect(ddahsDiff).toMatchSnapshot();
+    // import ddah with invalid assignment should throw error
+    expect(() =>
+        diffImport.ddahs(normalizedJsonDdahs, importDdahContextNoAssignment)
+    ).toThrow(Error);
 });
